+++ src/transport/drd_rdp_listener.h
@@ -11,10 +11,16 @@ G_BEGIN_DECLS
 #define DRD_TYPE_RDP_LISTENER (drd_rdp_listener_get_type())
 G_DECLARE_FINAL_TYPE(DrdRdpListener, drd_rdp_listener, DRD, RDP_LISTENER, GSocketService)
 
+typedef struct _DrdRdpSession DrdRdpSession;
+
 typedef gboolean (*DrdRdpListenerDelegateFunc)(DrdRdpListener *listener,
                                                GSocketConnection *connection,
                                                gpointer user_data,
                                                GError **error);
+typedef void (*DrdRdpListenerSessionFunc)(DrdRdpListener *listener,
+                                          DrdRdpSession *session,
+                                          GSocketConnection *connection,
+                                          gpointer user_data);
 
 DrdRdpListener *drd_rdp_listener_new(const gchar *bind_address,
                                      guint16 port,
@@ -34,5 +40,8 @@ void drd_rdp_listener_set_delegate(DrdRdpListener *self,
 gboolean drd_rdp_listener_adopt_connection(DrdRdpListener *self,
                                            GSocketConnection *connection,
                                            GError **error);
+void drd_rdp_listener_set_session_callback(DrdRdpListener *self,
+                                           DrdRdpListenerSessionFunc func,
+                                           gpointer user_data);
 
 G_END_DECLS